Move login_redirect_helper to login_app.
